Pull out all the classes to their own files and folders in order to 'simplify' project files.
