--- Rogue.js
@@ -1,978 +1,0 @@
-"use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.getPlayerTitle = exports.getPlayerName = exports.Enemy = exports.NPC = exports.Floor = exports.Dungeon = void 0;
-const ROT = require("rot-js");
-const fs = require("fs");
-const os = require("os");
-const startingTown_json_1 = __importDefault(require("./private/maps/startingTown.json"));
-const mapWidth = 75;
-const mapHeight = 40;
-const playerFOVRadius = 8;
-const previouslyExploredAlpha = 0.4;
-// #: Room wall
-// &: Hallway wall
-// %: Cave wall
-const WALL_TILES = ['#', '&', '%'];
-// .: Room floor
-// ,: Hallway floor
-// `: Cave floor
-const FLOOR_TILES = ['.', ',', '`'];
-const TILES_NONWALKABLE_BLOCKS_LIGHT = ['&', '#', '%', '♠', 'ƒ', '☺', '☻'];
-const TILES_NONWALKABLE_LIGHT_PASSES = ['Æ', 'æ', 'µ', '╤', '☼', ':', 'Φ', '═', '≈', '║', '♀', '¶', '╬', '₧', '╦', 'Ω'];
-const TILES_WALKABLE_BLOCKS_LIGHT = ['+', '⌠'];
-const TILES_WALKABLE_LIGHT_PASSES = [' ', '.', ',', '`', '"', '-', '<', '>', 'Θ', '╥', '~', '╣', '╠', '⌡', 'ⁿ', '░'];
-var sg = new ROT.StringGenerator(null);
-var nameFile = '';
-var malePlayerNameData;
-var femalePlayerNameData;
-var adjectiveData;
-var nicknameData;
-fs.readFile('./private/names/keeperrl_male_names.txt', function (err, data) {
-    malePlayerNameData = data.toString().split(os.EOL);
-});
-fs.readFile('./private/names/keeperrl_female_names.txt', function (err, data) {
-    femalePlayerNameData = data.toString().split(os.EOL);
-});
-fs.readFile('./private/names/adjectives.txt', function (err, data) {
-    adjectiveData = data.toString().split(os.EOL);
-});
-fs.readFile('./private/names/nicknames.txt', function (err, data) {
-    if (err) {
-        console.error(err);
-        return;
-    }
-    nicknameData = data.toString().split(os.EOL);
-});
-function getPlayerName() {
-    // Wait until name files are loaded in
-    while (!malePlayerNameData || !femalePlayerNameData) { }
-    // Decide which name file to use.
-    if (Math.round(Math.random())) {
-        var rnd = Math.floor(Math.random() * malePlayerNameData.length);
-        return malePlayerNameData[rnd];
-    }
-    else {
-        var rnd = Math.floor(Math.random() * femalePlayerNameData.length);
-        return femalePlayerNameData[rnd];
-    }
-}
-exports.getPlayerName = getPlayerName;
-function getPlayerTitle() {
-    // Wait until adjective and nickname files are loaded
-    while (!adjectiveData || !nicknameData) { }
-    // Choose title for player
-    var rnd = Math.floor(Math.random() * adjectiveData.length);
-    var adjective = adjectiveData[rnd];
-    rnd = Math.floor(Math.random() * nicknameData.length);
-    var nickname = nicknameData[rnd];
-    return ('the ' + adjective + ' ' + nickname);
-}
-exports.getPlayerTitle = getPlayerTitle;
-class Dungeon {
-    constructor(seed, floorNumber = 0, furthestFloor = floorNumber) {
-        this.floorNumber = floorNumber;
-        if (furthestFloor) {
-            this.furthestFloor = furthestFloor;
-        }
-        else {
-            this.furthestFloor = floorNumber;
-        }
-        this.player = new Player();
-        this.floors = [];
-        this.seed = seed;
-        ROT.RNG.setSeed(this.seed + this.floorNumber);
-        this.floors[this.floorNumber] = new Floor(mapWidth, mapHeight, this.floorNumber);
-    }
-    gotoFloor(floorNumber, upOrDown) {
-        if (floorNumber < 0)
-            return;
-        if (!this.floors[floorNumber]) {
-            // New level that needs to be created and stored.
-            // RNG for ROT set to initial seed + floorNumber
-            ROT.RNG.setSeed(this.seed + floorNumber);
-            this.floors[floorNumber] = new Floor(mapWidth, mapHeight, floorNumber, upOrDown === 'down');
-        }
-        this.floorNumber = floorNumber;
-    }
-    getCurrentFloor() {
-        return this.floors[this.floorNumber];
-    }
-    getFloorDataForClient({ includePlayerInfo = false } = {}) {
-        var floor = this.getCurrentFloor();
-        this.player.x = floor.playerX;
-        this.player.y = floor.playerY;
-        var returnObject = {
-            map: this.getClientMapData(),
-            playerX: floor.playerX,
-            playerY: floor.playerY,
-            player: null
-        };
-        if (includePlayerInfo) {
-            returnObject.player = this.player;
-        }
-        return returnObject;
-    }
-    useStairs(symbol) {
-        var cf = this.getCurrentFloor();
-        if (!symbol) {
-            symbol = cf.getMap()[cf.playerX + "," + cf.playerY];
-        }
-        else if (symbol !== cf.getMap()[cf.playerX + "," + cf.playerY]) {
-            return false;
-        }
-        if (symbol === ">") {
-            this.gotoFloor(this.floorNumber + 1);
-            return true;
-        }
-        else if (symbol === "<") {
-            this.gotoFloor(this.floorNumber - 1);
-            return true;
-        }
-        return false;
-    }
-    getPlayerInteractables() {
-        var nearbyInteractables = {};
-        var mapInteractables = this.getCurrentFloor().getInteractables();
-        // Check map for signs, people, stairs, doors, books, etc. in player's area
-        var playerInteractableReach = this.player.getNeighboringPositions();
-        Object.keys(mapInteractables).forEach(interactable => {
-            playerInteractableReach.forEach(tileLocation => {
-                var objectAtLocation = mapInteractables[interactable][tileLocation];
-                if (objectAtLocation) {
-                    // There is an interactable here
-                    if (!nearbyInteractables[interactable]) {
-                        nearbyInteractables[interactable] = {};
-                    }
-                    nearbyInteractables[interactable][tileLocation] = mapInteractables[interactable][tileLocation];
-                }
-            });
-        });
-        var mapNPCs = this.getCurrentFloor().getMap().npcs;
-        playerInteractableReach.forEach(tileLocation => {
-            if (mapNPCs[tileLocation]) {
-                if (!nearbyInteractables['npcs']) {
-                    nearbyInteractables['npcs'] = {};
-                }
-                nearbyInteractables['npcs'][tileLocation] = mapNPCs[tileLocation];
-            }
-        });
-        return nearbyInteractables;
-    }
-    /** mapAlphaValues
-     *
-     * @param x Player's X
-     * @param y Player's Y
-     * @returns an array of alpha values to be used on the map of the tiles
-     * Tiles seen by the player = 1
-     * Tiles the player hasn't seen = 0
-     * Tiles previously seen but not currently in FOV is a value between 0 and 1.
-     */
-    mapAlphaValues(x, y) {
-        if (x && y) {
-            return this.getCurrentFloor().updateFOV(x, y);
-        }
-        else {
-            return this.getCurrentFloor().updateFOV(this.getCurrentFloor().playerX, this.getCurrentFloor().playerY);
-        }
-    }
-    setAlphaValues(alpha) {
-        this.getCurrentFloor().mapExplored = alpha;
-    }
-    updatePlayerPosition(x, y) {
-        this.getCurrentFloor().setPlayerPosition(x, y);
-    }
-    getMap() {
-        this.getCurrentFloor().getMap();
-    }
-    getClientMapData() {
-        return {
-            tiles: this.getCurrentFloor().getMap().tiles,
-            fov: this.mapAlphaValues(),
-            npcs: this.getCurrentFloor().getMap().npcs
-        };
-    }
-}
-exports.Dungeon = Dungeon;
-class Floor {
-    constructor(width, height, levelNumber, spawnOnDownStairs = false) {
-        this.characters = [];
-        // Map Explored also contains the FOV for the player
-        this.mapExplored = {};
-        this.width = width;
-        this.height = height;
-        this.levelNumber = levelNumber;
-        if (this.levelNumber == 0) {
-            // TownData will get updated by players, so doors will be opened by previous players...
-            // This parse(stringify(TownData)) creates a fresh copy of the original town data for all players.
-            var mapData = JSON.parse(JSON.stringify(startingTown_json_1.default));
-            this.map = new DungeonMap(mapData);
-            this.playerX = 32;
-            this.playerY = 31;
-            return;
-        }
-        return;
-        // TODO: Update this to the new Map object.
-        if ((this.levelNumber) % 5 == 0) {
-            /* create a connected map where the player can reach all non-wall sections. Changed in new Rot.js version */
-            // var options = { connected: true }
-            var cellMap = new ROT.Map.Cellular(width, height);
-            /* cells with 1/2 probability */
-            cellMap.randomize(0.43);
-            /* make a few generations */
-            for (var i = 0; i < 5; i++)
-                cellMap.create();
-            for (var j = 0; j < height; j++) {
-                for (var i = 0; i < width; i++) {
-                    if (cellMap._map[i][j]) {
-                        // Cave floor
-                        this.map[i + ',' + j] = '`';
-                    }
-                    else {
-                        this.map[i + ',' + j] = ' ';
-                    }
-                }
-            }
-            var digger = new ROT.Map.Digger(width, height, { roomWidth: [3, 7], roomHeight: [3, 7], corridorLength: [2, 10], dugPercentage: 0.2 });
-            var digCallback = function (x, y, value) {
-                var key = x + "," + y;
-                // Only add room wall if there isn't a cave floor there.
-                if (value) {
-                    if (this.map[x + ',' + y] != '`') {
-                        this.map[key] = ' ';
-                    }
-                }
-                else {
-                    this.map[key] = ".";
-                } // Walls: ' ' Floor: '.'
-            };
-            var nullConnectCallback = (x, y, contents) => { return null; };
-            cellMap.connect(nullConnectCallback, 0, null);
-        }
-        else {
-            var digger = new ROT.Map.Digger(width, height, { roomWidth: [3, 7], roomHeight: [3, 7], corridorLength: [2, 4], dugPercentage: 0.24 });
-            //new ROT.Map.Uniform(width, height, {roomWidth: [3,6], roomHeight: [3,6], roomDugPercentage: 0.5});
-            var digCallback = function (x, y, value) {
-                var key = x + "," + y;
-                if (value) {
-                    this.map[key] = " ";
-                }
-                else {
-                    this.map[key] = ".";
-                } // Walls: ' ' Floor: '.'
-            };
-        }
-        digger.create(digCallback.bind(this));
-        this.rooms = digger.getRooms();
-        var worldMap = this.map;
-        var drawDoor = function (x, y) {
-            worldMap[x + "," + y] = "+";
-        };
-        for (var i = 0; i < this.rooms.length; i++) {
-            var room = this.rooms[i];
-            room.getDoors(drawDoor);
-        }
-        for (var j = 0; j < this.height; j++) {
-            for (var i = 0; i < this.width; i++) {
-                if (this.map[i + "," + j] === ".") {
-                    if (!(this.inRoom(i, j))) {
-                        this.map[i + "," + j] = ",";
-                    }
-                }
-                this.mapExplored[i + "," + j] = 0;
-            }
-        }
-        for (var y = 0; y < this.height; y++) {
-            for (var x = 0; x < this.width; x++) {
-                if (this.map[x + "," + y] === " ") {
-                    var surroundingTiles = [this.map[x + "," + (y - 1)], this.map[x + "," + (y + 1)], this.map[(x - 1) + "," + y], this.map[(x + 1) + "," + y],
-                        this.map[(x + 1) + "," + (y - 1)], this.map[(x + 1) + "," + (y + 1)], this.map[(x - 1) + "," + (y - 1)], this.map[(x - 1) + "," + (y + 1)]].join('').trim();
-                    if (surroundingTiles.includes(".")) {
-                        // If a blank tile has a floor tile around it, then the blank tile needs to be a wall
-                        this.map[x + "," + y] = "#";
-                    }
-                    else if (surroundingTiles.includes(",")) {
-                        // If the surrounding tile includes a hallway or a cave floor, add a cave wall.
-                        this.map[x + "," + y] = "&";
-                    }
-                    else if (surroundingTiles.includes("`")) {
-                        this.map[x + "," + y] = "%";
-                    }
-                }
-            }
-        }
-        //var playerStartRoom = this.rooms[Math.floor(Math.random() * (this.rooms.length-1))];
-        if (this.levelNumber != 0) {
-            this.map[this.rooms[0].getCenter()[0] + "," + this.rooms[0].getCenter()[1]] = "<";
-        }
-        //var stairsDownRoom = playerStartRoom;
-        //while (stairsDownRoom == playerStartRoom) { stairsDownRoom = this.rooms[Math.floor(Math.random() * (this.rooms.length-1))]; }
-        this.map[this.rooms[this.rooms.length - 1].getCenter()[0] + "," + this.rooms[this.rooms.length - 1].getCenter()[1]] = ">";
-        var roomID = 0;
-        if (spawnOnDownStairs) {
-            roomID = this.rooms.length - 1;
-        }
-        this.playerX = this.rooms[roomID].getCenter()[0];
-        this.playerY = this.rooms[roomID].getCenter()[1];
-        // Initiate Enemies and Dijkstra path-finding passed function
-        //this.placeEnemies();
-        this.dijkstra = new ROT.Path.Dijkstra(this.playerX, this.playerY, function (x, y) {
-            if (this.map) {
-                return ((this.map[x + ',' + y] === '.') || (this.map[x + ',' + y] === ',') || (this.map[x + ',' + y] === '-') || (this.map[x + ',' + y] === '+') || (this.map[x + ',' + y] === '`'));
-            }
-            else {
-                return false;
-            }
-        }, null);
-    }
-    updateFOV(pX, pY) {
-        var localMap = this.map.getAsciiTiles();
-        var localMapExplored = this.mapExplored;
-        var previousMapExplored = {};
-        Object.keys(this.mapExplored).forEach(key => {
-            previousMapExplored[key] = this.mapExplored[key];
-        });
-        // Player's field-of-view light input callback
-        var lightPasses = function (x, y) {
-            var key = x + "," + y;
-            if (key in localMap) {
-                return (TILES_NONWALKABLE_LIGHT_PASSES.includes(localMap[key]) || TILES_WALKABLE_LIGHT_PASSES.includes(localMap[key]));
-            }
-            // If the tile is undefined, return true.
-            return true;
-        };
-        var fov = new ROT.FOV.PreciseShadowcasting(lightPasses);
-        // Output callback for player's field-of-view
-        fov.compute(pX, pY, playerFOVRadius, function (x, y, r, visibility) {
-            var ch = (r ? "" : "@");
-            localMapExplored[x + "," + y] = 2;
-        });
-        for (var j = 0; j < this.height; j++) {
-            for (var i = 0; i < this.width; i++) {
-                var tileAlpha = this.mapExplored[i + "," + j];
-                if (tileAlpha == 1) {
-                    tileAlpha = previouslyExploredAlpha;
-                }
-                else if (tileAlpha == 2) {
-                    tileAlpha = 1;
-                }
-                else if (tileAlpha != previouslyExploredAlpha) {
-                    tileAlpha = 0;
-                }
-                this.mapExplored[i + "," + j] = tileAlpha;
-            }
-        }
-        if (this.diffMapExplored == null) {
-            this.diffMapExplored = this.mapExplored;
-        }
-        else {
-            // Calculate which tiles had updated FOV values
-            var diff = {};
-            Object.keys(this.mapExplored).forEach(pos => {
-                if (this.mapExplored[pos] != previousMapExplored[pos]) {
-                    diff[pos] = this.mapExplored[pos];
-                }
-            });
-            if (Object.keys(diff).length > 0) {
-                // Only update when the diff is different to prevent deleting data not sent to player.
-                this.diffMapExplored = diff;
-            }
-            else {
-                // FOV values are the same from previous turn.
-                this.diffMapExplored = {};
-            }
-        }
-        return this.mapExplored;
-    }
-    getDiffMapExplored() {
-        return this.diffMapExplored;
-    }
-    getMap() {
-        return this.map;
-    }
-    getPlayerPosition() {
-        return { x: this.playerX, y: this.playerY };
-    }
-    inRoom(x, y) {
-        for (var i = 0; i < this.rooms.length; i++) {
-            var room = this.rooms[i];
-            if (x >= room.getLeft() && x <= room.getRight() && y >= room.getTop() && y <= room.getBottom()) {
-                return i + 1;
-            }
-        }
-        return 0;
-    }
-    setPlayerPosition(x, y) {
-        this.playerX = x;
-        this.playerY = y;
-        if (this.map.getTile(x, y)) {
-            if (this.map.getTile(x, y).ascii === '+') {
-                this.map.setTileAscii(x, y, '-');
-                ;
-            }
-            else if (this.map.getTile(x, y).ascii === '╣') {
-                this.map.setTileAscii(x, y, '╠');
-            }
-        }
-    }
-    getExploredMap() {
-        return this.mapExplored;
-    }
-    openDoor(x, y) {
-        if (this.map[x + "," + y] === '+') {
-            this.map[x + "," + y] = '-';
-            return true;
-        }
-        else {
-            return false;
-        }
-    }
-    canWalk(x, y) {
-        if (x > this.width || y > this.height || x < 0 || y < 0) {
-            return false;
-        }
-        switch (this.map[x + "," + y]) {
-            // Hallway Wall
-            case '&':
-            // Room Wall
-            case '#':
-            // Cave Wall
-            case '%':
-                return false;
-            default:
-                return true;
-        }
-    }
-    /*
-    placeEnemies() {
-        var enemyStartRoom, enemyX, enemyY, attempts = 0;
-        // TODO: Improve enemy placement
-        // Attempts is to prevent infinite loops that might happen during test. Will place enemies different in final implementation
-        do {
-            enemyStartRoom = this.rooms[Math.floor(Math.random() * (this.rooms.length - 1))];
-            enemyX = enemyStartRoom.getCenter()[0];
-            enemyY = enemyStartRoom.getCenter()[1];
-            attempts++;
-        } while (this.inRoom(enemyX, enemyY) == this.inRoom(this.playerX, this.playerY) && attempts < 3);
-        this.enemies.push(new Enemy('goblin', enemyX, enemyY));
-    }
-    getEnemies() {
-        return this.enemies;
-    }
-    getEnemyAt(x, y) {
-        for (var i = 0; i < this.enemies.length; i++) {
-            var enemy = this.enemies[i];
-
-            if (enemy.x == x && enemy.y == y) {
-                return enemy;
-            }
-        }
-        return null;
-    }*/
-    generateTileData() {
-        // Instead of placing individual tiles, store all tile sprites together in an array
-        // in the layout of map[x+","+y] with the key being tile coordinates. This will allow
-        // easier updating of an individual tiles alpha value.
-        /*
-
-        Logic for new tile names.
-        Create dictionary if possible. Would need to allow wildcards or subsets
-        Not sure if better to create custom class, database, or rely on other methods.
-        Dictionary string would be 8 digits long that represent all surround tiles (9x9 grid with tile getting name for in center)
-        |0|1|2|
-        |3|X|4|
-        |5|6|7|
-
-        TODO: Rotation of tiles that require it will need to have their pivot x and y values set to 16, middle of sprite
-
-        FLOOR DECORATIONS
-        wall_borders floor tiles (low percentage):
-            Check wall and door location
-            | wall locations  | tile name       |
-            | wall (x-1)(y+1) | floor_border_NW |
-            | wall ( x )(y+1) | floor_border_N  |
-            | wall (x+1)(y+1) | floor_border_NE |
-            | wall (x-1)( y ) | floor_border_W  |
-            | wall (x+1)( y ) | floor_border_E  |
-            | wall (x-1)(y-1) | floor_border_SW |
-            | wall ( x )(y-1) | floor_border_S  |
-            | wall (x+1)(y-1) | floor_border_SE |
-
-            floor_border_ + join()
-                if wall(y+1) N
-                if wall(y-1) S
-                if wall(x+1) E
-                if wall(x-1) W
-
-        if no surrounding walls (use NSEW joined screen from above)
-            Choose secondary floor texture
-                floor_1-8
-
-        if hall_way:
-            cave_floor_1-3 and prioritize lower numbers
-
-        if cave_floor:
-            cave_floor_2-6 and prioritize higher numbers
-
-
-            TODO: Modify the map creation and tiles to add cave walls.
-            % is the cave floor tiles.
-                     #####  %%%%%#####&&
-                    #...# %%`````...+,&
-                    #...# %``````...`,&
-                    #...# %%`````...`,&
-                    #...#  %%%%%`````,&
-                &&#...#&&&&&&&&&```,&
-            %%%&,+...+,,,,,,,,,&%&,&
-            %%``&,#...#&&######+##&,&
-            %```&,#####  #.......+,,&
-        %%%  %``%&,&      #.......#&&&
-        %%`%% %``%&,&      #.......#
-        %```% %``%&,&      #.......#
-    %%```% %%%%&,&    %%#.......#
-    %```%%     &,&   %%``.......#
-    %```%      &,&   %```+#######
-    %```%      &,&   %```,&%
-    %%%%%#######+#####%``,`%
-            #....#......#%%`,&%
-            #....#......# %&,&
-    %%%%%  #....+...<..#  &,&
-    %```%  ######......#  &,&
-    %```%       #......#  &,&
-    %%`%%   %%% ########  &,&   #####
-    %%%    %`%%       ####+#   #...#
-            %``%       #....#%  #.>.#
-            %`%%       #....`%  #...#
-            %%%        #....#%#####+#
-                    %#....# #.....#
-                    %%``+### #.....#
-                    %```,&%  #.....#
-        %%%        %```,`%% #.....#
-        %%`%%       %%``,``% #.....#
-        %```%        %%`,`&&&##+####
-        %```%         %&,,,,,,,,&
-        %%%%%          &&&&&&&&&&
-                                                    
-
-
-        */
-        var mapTileData = {};
-        var tileName;
-        var mapData = ' ';
-        var localMap = this.map;
-        for (let y = 0; y < mapHeight; y++) {
-            for (let x = 0; x < mapWidth; x++) {
-                mapData = this.map[x + "," + y];
-                switch (mapData) {
-                    case '&':
-                        /*
-                        This code block is trying to figure out making caves walls that work like the connecting room walls.
-
-                        // This .join('').trim() will return an empty string if this is a map tile in the void.
-                        // TODO: This doesn't work when one of these tiles is null!
-                        function getMapTile(x, y) {
-                            var tile = localMap[x+","+y];
-                            if (tile)
-                                return tile
-                            return " "
-                        }
-                        var surroundingTiles = [getMapTile(x,y-1), getMapTile(x,y+1), getMapTile(x-1,y), getMapTile(x+1,y),
-                                            getMapTile(x+1,y-1), getMapTile(x+1,y+1), getMapTile(x-1,y-1), getMapTile(x-1,y+1)].join('');
-                        
-                        
-                        var countCaveFloors = (surroundingTiles.substring(0,4).match(/,/g) || []).length;
-                        if (countCaveFloors == 0) {
-                            // If the cave floor doesn't have an immediate neighbor, then it is a corner piece so check diagonal tiles.
-                            countCaveFloors = (surroundingTiles.substring(4).match(/,/g) || []).length;
-                        }
-                        var surroundingWalls = "";
-                        if (countCaveFloors == 1) {
-                            // This will give location of the closest cave floor tile that the wall should be positioned around.
-                            var floorLocation = surroundingTiles.indexOf(",");
-
-                            var caveWallSpriteFromFloor = {
-                                0: 'EW',
-                                1: 'EW',
-                                2: 'NS',
-                                3: 'NS',
-                                4: 'SW',
-                                5: 'NW',
-                                6: 'SE',
-                                7: 'NE'
-                            }
-
-                            tileData = "cave_wall_" + caveWallSpriteFromFloor[floorLocation];
-                            //console.log(`Cave wall tile: ${tileData}`);
-                        } else {
-                            if (surroundingTiles[0] == ',') {
-                                surroundingWalls += "N"
-                            }
-                            if (surroundingTiles[1] == ',') {
-                                surroundingWalls += "S"
-                            }
-                            if (surroundingTiles[3] == ',') {
-                                surroundingWalls += "E"
-                            }
-                            if (surroundingTiles[2] == ',') {
-                                surroundingWalls += "W"
-                            }
-
-                            tileData = "cave_wall_" + surroundingWalls;
-                        }
-
-
-                        // TODO: Create an object with all the valid floor names and check against it.
-                        var validTiles = ["cave_wall_NS", "cave_wall_EW", "cave_wall_NW", "cave_wall_NE", "cave_wall_SE", "cave_wall_SW", "cave_wall_NSE", "cave_wall_SEW", "cave_wall_NSW", "cave_wall_NEW", "cave_wall_NSEW"]
-
-                        if (!tileData || validTiles.indexOf(tileData) == -1) {
-                            console.log(`map position ${x},${y} has invalid tile name: ${tileData}`);
-                            tileData = "player_defeated"
-                        }
-                        */
-                        var randomNumberBasedOnMapLocation = this.chooseTexture(x, y, 100);
-                        if (randomNumberBasedOnMapLocation < 90) {
-                            tileName = "hallway_wall_1";
-                        }
-                        else {
-                            tileName = "hallway_wall_2";
-                        }
-                        break;
-                    case '.':
-                        tileName = "room_floor_" + this.chooseTexture(x, y, 7);
-                        break;
-                    case ',':
-                        tileName = "cave_floor_" + this.chooseTexture(x, y, 4);
-                        break;
-                    case '>':
-                        tileName = "stairs_down";
-                        break;
-                    case '<':
-                        tileName = "stairs_up";
-                        break;
-                    case '+':
-                        tileName = "door";
-                        break;
-                    case '-':
-                        tileName = "doorOpen";
-                        break;
-                    case '`':
-                        // Cave floor
-                        tileName = "cave";
-                        break;
-                    case '#':
-                        // Surround Tiles is an array of the results from checking if the tiles surrounding a wall are in rooms or not.
-                        // Each direction is used twice so wanted to reduce the number of calls to inRoom
-                        // Order is in 0)N, 1)S, 2)E, 3)W, 4)NE, 5)NW, 6)SE, 7)SW
-                        var surroundingTilesInRooms = [this.inRoom(x, (y - 1)), this.inRoom(x, (y + 1)), this.inRoom((x + 1), y), this.inRoom((x - 1), y),
-                            this.inRoom((x + 1), (y - 1)), this.inRoom((x - 1), (y - 1)), this.inRoom((x + 1), (y + 1)), this.inRoom((x - 1), (y + 1))];
-                        var name = "";
-                        if (!surroundingTilesInRooms[0] && (surroundingTilesInRooms[2] + surroundingTilesInRooms[3] + surroundingTilesInRooms[4] + surroundingTilesInRooms[5])) {
-                            // If there is a room to the East or West of the wall, it's a wall that connects on the North and South side.
-                            name += "N";
-                        }
-                        else {
-                            name += "_";
-                        }
-                        if (!surroundingTilesInRooms[1] && (surroundingTilesInRooms[2] + surroundingTilesInRooms[3] + surroundingTilesInRooms[6] + surroundingTilesInRooms[7])) {
-                            // If there is a room to the East or West of the wall, it's a wall that connects on the North and South side.
-                            name += "S";
-                        }
-                        else {
-                            name += "_";
-                        }
-                        if (!surroundingTilesInRooms[2] && (surroundingTilesInRooms[0] + surroundingTilesInRooms[1] + surroundingTilesInRooms[4] + surroundingTilesInRooms[6])) {
-                            // If there is a room to the East or West of the wall, it's a wall that connects on the North and South side.
-                            name += "E";
-                        }
-                        else {
-                            name += "_";
-                        }
-                        if (!surroundingTilesInRooms[3] && (surroundingTilesInRooms[0] + surroundingTilesInRooms[1] + surroundingTilesInRooms[5] + surroundingTilesInRooms[7])) {
-                            // If there is a room to the East or West of the wall, it's a wall that connects on the North and South side.
-                            name += "W";
-                        }
-                        else {
-                            name += "_";
-                        }
-                        tileName = name;
-                        break;
-                    case '%':
-                        function getMapTile(x, y) {
-                            var tile = localMap[x + "," + y];
-                            if (tile)
-                                return tile;
-                            return " ";
-                        }
-                        var surroundingTiles = [getMapTile(x, y - 1), getMapTile(x, y + 1), getMapTile(x + 1, y), getMapTile(x - 1, y)];
-                        var connectedWalls = ['', '', '', ''];
-                        // Trying to find the best wall to connect cave wall sprites.
-                        // First check for any nearby cave walls that this tile should connect to
-                        if (surroundingTiles[0] == '%') {
-                            connectedWalls[0] = 'N';
-                        }
-                        if (surroundingTiles[1] == '%') {
-                            connectedWalls[1] = 'S';
-                        }
-                        if (surroundingTiles[2] == '%') {
-                            connectedWalls[2] = 'E';
-                        }
-                        if (surroundingTiles[3] == '%') {
-                            connectedWalls[3] = 'W';
-                        }
-                        // If the tile could only connect to 1 cave wall, try to see if there are other wall tiles nearby to connect to
-                        if (connectedWalls.join('').trim().length <= 1) {
-                            if (WALL_TILES.includes(surroundingTiles[0])) {
-                                connectedWalls[0] = 'N';
-                            }
-                            if (WALL_TILES.includes(surroundingTiles[1])) {
-                                connectedWalls[1] = 'S';
-                            }
-                            if (WALL_TILES.includes(surroundingTiles[2])) {
-                                connectedWalls[2] = 'E';
-                            }
-                            if (WALL_TILES.includes(surroundingTiles[3])) {
-                                connectedWalls[3] = 'W';
-                            }
-                        }
-                        tileName = "cave_wall_" + connectedWalls.join('').trim();
-                        var validTiles = [
-                            "cave_wall_NS",
-                            "cave_wall_EW",
-                            "cave_wall_SE",
-                            "cave_wall_SW",
-                            "cave_wall_NW",
-                            "cave_wall_NE",
-                            "cave_wall_NEW",
-                            "cave_wall_NSE",
-                            "cave_wall_SEW",
-                            "cave_wall_NSW",
-                            "cave_wall_NSEW"
-                        ];
-                        if (!validTiles.includes(tileName)) {
-                            console.log(`ERROR: Invalid tile at ${x},${y}: ${tileName}`);
-                            tileName = "door_floor_wear_1";
-                        }
-                        break;
-                    default:
-                        tileName = mapData;
-                }
-                if (tileName) {
-                    // Only add the tile if it contains data.
-                    mapTileData[x + "," + y] = tileName;
-                }
-            }
-        }
-        return mapTileData;
-    }
-    chooseTexture(x, y, z) {
-        // https://stackoverflow.com/questions/12964279/whats-the-origin-of-this-glsl-rand-one-liner
-        // x and y in the below equation should be divided by total height or width so that way x and y
-        // will always be between 0 and 1.
-        var number = (Math.sin((x / this.width) * 12.9898 + (y / this.height) * 78.233) * 43758.5453) % 1;
-        //console.log(number);
-        var newNumber = Math.abs(Math.floor(number * z)) + 1;
-        if (newNumber > z) {
-            return newNumber - z;
-        }
-        else {
-            return newNumber;
-        }
-    }
-    getInteractables() {
-        return this.map.interactables;
-    }
-}
-exports.Floor = Floor;
-class Entity {
-    constructor(x, y) {
-        this.x = x;
-        this.y = y;
-    }
-    getPositionString() {
-        return `${this.x},${this.y}`;
-    }
-    setPosition(x, y) {
-        this.x = x;
-        this.y = y;
-    }
-    getNeighboringPositions() {
-        // Used to get the tiles x-1, x, x+1 and y-1, y, y+1 easily using some loops
-        var touchingTilesReach = [-1, 0, 1];
-        var neighboringTilesString = [`${this.x},${this.y}`];
-        touchingTilesReach.forEach(xMod => {
-            touchingTilesReach.forEach(yMod => {
-                if (Math.abs(xMod) != Math.abs(yMod)) {
-                    var tileX = this.x + xMod;
-                    var tileY = this.y + yMod;
-                    if (tileX >= 0 && tileX < mapWidth &&
-                        tileY >= 0 && tileY < mapHeight) {
-                        neighboringTilesString.push(`${tileX},${tileY}`);
-                    }
-                }
-            });
-        });
-        return neighboringTilesString;
-    }
-    // This function covers tiles diagonal from entity
-    getAllNeighboringPositions() {
-        // Used to get the tiles x-1, x, x+1 and y-1, y, y+1 easily using some loops
-        var touchingTilesReach = [-1, 0, 1];
-        var neighboringTilesString = [];
-        touchingTilesReach.forEach(xMod => {
-            touchingTilesReach.forEach(yMod => {
-                var tileX = this.x + xMod;
-                var tileY = this.y + yMod;
-                if (tileX >= 0 && tileX < mapWidth &&
-                    tileY >= 0 && tileY < mapHeight) {
-                    neighboringTilesString.push(`${tileX},${tileY}`);
-                }
-            });
-        });
-        return neighboringTilesString;
-    }
-}
-class Player extends Entity {
-    constructor(name = null) {
-        super(0, 0);
-        this.health = 10;
-        this.attack = [1, 2];
-        if (name) {
-            this.name = name;
-        }
-        else {
-            this.name = getPlayerName();
-        }
-        this.title = getPlayerTitle();
-    }
-}
-class Enemy extends Entity {
-    constructor(name, x, y) {
-        super(x, y);
-        this.char = "";
-        this.loot = [];
-        this.description = "";
-        this.path = [];
-        this.name = name;
-        this.x = x;
-        this.y = y;
-        this.health = 0;
-        this.attack = [1];
-        this.accuracy = 0;
-        switch (name) {
-            case 'goblin':
-                this.health = 4;
-                this.maxHealth = this.health;
-                this.attack = [1];
-                this.accuracy = 0.8;
-                this.char = "g";
-                this.description = "It a goblin.";
-                break;
-        }
-    }
-    calculateMove(playerX, playerY, map) {
-        var dijkstra = new ROT.Path.Dijkstra(playerX, playerY, function (x, y) {
-            if (map) {
-                return ((map[x + ',' + y] === '.') || (map[x + ',' + y] === ',') || (map[x + ',' + y] === '-') || (map[x + ',' + y] === '+') || (map[x + ',' + y] === '`'));
-            }
-            else {
-                return false;
-            }
-        }, { topology: 4 });
-        var localPath = this.path;
-        dijkstra.compute(this.x, this.y, function (x, y) {
-            // For testing, have the alpha of all the tiles in the enemies path should have the alpha value of 1.
-            localPath.push({ x: x, y: y });
-        });
-        // Remove the first spot in path because it contains the enemy's current location. We want the first spot to be where the enemy's next turn should be.
-        this.path.shift();
-        return this.path;
-    }
-}
-exports.Enemy = Enemy;
-class NPC extends Entity {
-    constructor(npcInfo) {
-        super(npcInfo.x, npcInfo.y);
-        this.name = npcInfo.name;
-        this.ascii = npcInfo.symbol;
-        this.sprite = npcInfo.sprite;
-        this.hostile = npcInfo.hostile;
-    }
-}
-exports.NPC = NPC;
-class DungeonMap {
-    constructor(townData) {
-        this.tiles = {};
-        var tilesData = {};
-        if (townData.hasOwnProperty("tiles")) {
-            tilesData = townData.tiles;
-        }
-        Object.keys(tilesData).forEach(tileLocation => {
-            tilesData[tileLocation] = new Tile(tilesData[tileLocation]);
-        });
-        this.tiles = tilesData;
-        if (townData.hasOwnProperty("npcs")) {
-            this.npcs = townData.npcs;
-        }
-        if (townData.hasOwnProperty("interactables")) {
-            this.interactables = townData.interactables;
-        }
-    }
-    getAsciiTiles() {
-        var asciiTiles = {};
-        Object.keys(this.tiles).forEach(tileLocation => {
-            if (this.tiles[tileLocation].ascii) {
-                asciiTiles[tileLocation] = this.tiles[tileLocation].ascii;
-            }
-        });
-        return asciiTiles;
-    }
-    getTile(x, y) {
-        return this.tiles[x + ',' + y];
-    }
-    setTileAscii(x, y, value) {
-        this.tiles[x + ',' + y] = value;
-    }
-}
-class Tile {
-    constructor(tileData) {
-        if (tileData.ascii) {
-            this.ascii = tileData.ascii;
-        }
-        if (tileData.sprite) {
-            this.sprite = tileData.sprite;
-        }
-    }
-}
-class TownRoom {
-    constructor() {
-        this.description = "";
-    }
-}
-class Interactable {
-    constructor() {
-        this.interacted = false;
-        this.description = "";
-    }
-}
-class Sign extends Interactable {
-    constructor() {
-        super();
-        this.text = "";
-    }
-}
-class Book extends Interactable {
-    constructor() {
-        super();
-        this.title = "";
-        this.author = "";
-    }
-}
-class Door extends Interactable {
-    constructor() {
-        super(...arguments);
-        this.locked = false;
-    }
-}
-//# sourceMappingURL=Rogue.js.map
