+++ app.ts
@@ -1,13 +1,13 @@
-﻿import express = require('express');
+﻿require('module-alias/register')
+import express = require('express');
 import { createServer } from "http";
 import { Server, Socket } from "socket.io";
 import { v4 as uuidv4 } from 'uuid';
 // import sqlite3 = require('sqlite3');
-import Rogue = require('./Rogue');
+import Rogue = require('./src/Rogue');
 
 // Local port the server will listen to connections on.
 const PORT = 1337;
-
 
 let app = express();
 const server = createServer(app);
@@ -25,10 +25,7 @@ app.get('/', function (req, res) {
     res.sendFile(__dirname + '/public/index.html');
 });
 
-
-
-
-io.on('connection', function (socket) {
+io.on('connection', function (socket: Socket) {
     console.log('a user connected');
     var uuid;
     var dungeon;
@@ -179,15 +176,14 @@ io.on('connection', function (socket) {
     });
 });
 
-process.stdout.write(
-    String.fromCharCode(27) + "]0;" + "Labyrinthine Flight Server" + String.fromCharCode(7)
-);
-
 server.listen(PORT, function () {
     console.log('Listening on port ' + PORT + '.');
 });
 
-process.on('SIGINT', function () {
+var shutdown = (signal = null) => {
     console.log('Shuting down...');
+    server.close()
     process.exit(0);
-});
+}
+
+process.on('SIGINT', shutdown);
