+++ package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "debug": "^2.2.0",
     "express": "^4.17.1",
+    "module-alias": "^2.2.2",
     "rot-js": "^2.1.5",
     "socket.io": "^4.1.2",
     "sqlite3": "^5.0.2",
@@ -29,5 +30,11 @@
     "@types/node": "^8.0.14",
     "@types/serve-static": "^1.7.32",
     "typescript": "^3.9.10"
+  },
+  "_moduleAliases": {
+    "@world": "./src/world",
+    "@entities": "./src/entities",
+    "@maps": "./src/maps",
+    "@utilities": "./src/utilities"
   }
 }
