+++ src/Rogue.ts
@@ -1,0 +1,1 @@
+﻿export { default as Dungeon } from '@world/dungeon/Dungeon'
