+++ src/utilities/PlayerDetails.ts
@@ -1,0 +1,16 @@
+export default class PlayerDetails {
+    malePlayerNameData: string[]
+    femalePlayerNameData: string[]
+    adjectiveData: string[]
+    nicknameData: string[]
+
+    constructor() {}
+
+    getPlayerName() {
+        return "Tester"
+    }
+
+    getPlayerTitle() {
+        return 'the In-Dev Adventurer';
+    }
+}
