+++ src/utilities/names/adjectives.txt
