+++ src/world/Town.ts
@@ -1,0 +1,10 @@
+class TownRoom {
+    name: string;
+    x1: number;
+    y1: number;
+    x2: number;
+    y2: number;
+    description: string = "";
+}
+
+export { TownRoom }
