+++ tsconfig.json
@@ -1,11 +1,29 @@
 {
   "compilerOptions": {
     "module": "commonjs",
-    "target": "es6",
-    "lib": ["es6", "dom"],
+    "lib": [
+      "es6",
+      "dom"
+    ],
+    "moduleResolution": "node",
     "sourceMap": true,
     "resolveJsonModule": true,
-    "esModuleInterop": true
+    "esModuleInterop": true,
+    "baseUrl": "./src",
+    "paths": {
+      "@world/*": [
+        "world/*"
+      ],
+      "@entities/*": [
+        "entities/*"
+      ],
+      "@maps/*": [
+        "maps/*"
+      ],
+      "@utilities/*": [
+        "utilities/*"
+      ],
+    }
   },
   "exclude": [
     "node_modules"
